karchivetest: testReadTar: fix user/group checks for dirs
When checking user and group of directories created automatically
(i.e. created when adding a file), compare them with the system user
name and group name, as they are the attributes set for them (see
KArchive::rootDir() and KArchive::findOrCreate() for reference).

Instead, in the local file case, keep checking with the reference
file attributes, but without falling back on the system user name and
group name (KArchive::addLocalFile() does not do such fallback).