Handle multiple gzip streams
If kgzipfilter notices that zlib didn't read all the data, it tries to
re-init the stream and read the rest of the buffer. This is tested by
the unit test from Sune. The case where there's more than two streams
available in the current buffer is tested in a unit test added
separately.

If the split between the streams falls right between two buffers, we
need KCompressionDevice to notice that there's data left and try to
continue decompressing. This is easy to test by setting BUFFER_SIZE to
the size of the first stream in the unit test data (28 bytes).

REVIEW: 128369
BUG: 232843