Revert "Return KFilterBase::End if KGzipFilter::uncompress_noop() finished reading"
This breaks rendering various SVG icons in Plasma. Namely ones that are "minified" and consist
of one long line of text, for example the new "kolf" icon and even "systemsettings".

Plasma's icon item loads SVG icons, even ones from an icon theme rather than a Plasma theme,
to apply Plasma theme colors to it. As the Plasma Svg Renderer is from a time before that was done
it assumes icons are SVGZ and alwys runs them through KCompressionDevice, which seems to no-op
when the file isn't actually compressed. However, this bug still struck as
compressionDevice.load() only returned 8192 bytes (one BUFFER_SIZE) and the resulting SVG data
was truncated and as such invalid and failed to load.

This reverts commit 96083005e90e1f4b0bfe8aa6344e60c1f1cdcc77.

CCMAIL: chrisito@gmx.de