Fix memory leak when reading some tar files
Summary:
Not sure if they are even malformed (don't know enough about the tar format), but oss-fuzz has produced a file where there's files and then there's the special . file (the root).

Since the files prior to finding "." need to be child of something, we already created a rootdir, so when we find "." we need to delete the existing one.

Reviewers: apol

Reviewed By: apol

Subscribers: apol, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D20644