Rewrite KZip file parsing
The old file parser was reading the file in order looking for the
local header and central directory header markers, but that breaks when
those markers are part of the data of the files inside the zip file
itself.

The correct way is to scan for the end of central directory record
signature and then from there use the offsets to jump to the rest of the
data.

BUGS: 450597