kzip: Fix misdetection of nested signatures, handle Zip64 data descriptors
In case a zip file entry has an undetermined size in the local header
the end of the binary data is marked by a data descriptor, with
signature 'PK\7\8'.

This will occasionally fail for two different reasons:
1. Compressed data randomly matching the signature
2. Stored, nested ZIP files.

In the first case, almost any sanity check for the following data
will do.

In the second case, the 'compressed size' field of the data descriptor
will not match the current position difference. Only for the data
descriptor of the outer ZIP file (start position + uncompressed size ==
current position) will match.

Also fix the size of the Zip64 data descriptors, and skip to the next
header correctly. Zip64 data descriptors are marked by a "needed version"
of 45 or later in the preceding local header.

The header seeking is mostly rewritten. Instead of seeking backwards
and forwards, and reading one character at a time from the underlying
QIODevice, a larger block is read using peek(). Scanning through
the block for a 'PK' signature is significantly faster than the
one-byte-at-a-time approach.

BUG: 450597