Skip two bytes on invalid header signature in `seekToNextHeaderToken()`
`seekAnyHeader()` advances the stream to the start of next potential
header but may not move if the stream is already positioned at a header
candidate. When encountering a token that does not match an expected
signature (i.e, PK\x03\x04 or PK\x01\x02) in `seekToNextHeaderToken()`,
manually advance the stream by two bytes to avoid re-detecting the same
invalid token in the next iteration.