K7Zip: Avoid recurrent temporary buffer allocation, fix shadowing
The method/codec ID is stored as a variable lenght, MSB first value.
It can be read from the back or the front, as long as each byte is
shifted correctly before summing it.

The only purpose of the buffer is to (eventually) change the signedness
from char to unsigned char, which can be done when reading each byte.

Also fixes shadowing the outer loop variable `i`.